poker-pipeline/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
scipy==1.10.1


